packages/create_myte/templates/template-flask-fullstack/requirements.txt
alembic==1.13.1
ansicon==1.89.0
autopep8==2.0.4
blessed==1.20.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
editor==1.6.6
filelock==3.13.1
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
inquirer==3.2.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.2.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
myte==1.0.0b2
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
python-dotenv==1.0.1
readchar==4.0.5
rich==13.7.0
runs==1.2.2
six==1.16.0
SQLAlchemy==2.0.27
typer==0.9.0
typing_extensions==4.9.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
xmod==1.8.1
